Created Classes
Created classes with some stats to go with each unique class. Basic implementation of the classes giving each a health stat, an attack damage stat, and an ability power stat which all change accordingly to match each class. Implemented a testing feature in the main just to make sure everything was working properly and it was.