Level Up System
Implemented leveling up system on a very basic scale. Came up with the calculation that allows the classes to level up and the exp required to level up increase as the character levels up. Also implemented stat scaling for characters so that each characters stats scale as they level up. Implemented multiple functions like getLevel  or getEXPToNextLevel that just allows me to print out the level or xp needed to the terminal so the user can see the characters information and their stats improve.