Cleanup
Changed a lot of stuff in the code but it was mainly all cleanup, condensed all of the class files into a single player character file where everything is delegated. Also modularized everything so that in the future if I want to add new classes or even new stats like mana or energy I can without having to implement it all over again. Created welltype for that reason specifically. Also created macros for leveling up and constructing a classes stats so that code looks a little cleaner and allows for multiple characters/classes to be created with ease.